+++ app/registration/schemas.py
@@ -1,0 +1,10 @@
+from app.users.schemas import UserCreate
+
+
+class RegistrationRequest(UserCreate):
+    pass
+
+
+class RegistrationResponse():
+    access_token: str
+
